requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
pyyaml==5.3.1
